+++ Editor/Artifice_CustomAttributeDrawers/CustomAttributeDrawers_Groups/Artifice_CustomAttributeDrawer_FoldoutGroupAttribute.cs
@@ -5,7 +5,7 @@ namespace ArtificeToolkit.Editor.Artifice_CustomAttributeDrawers.CustomAttribute
 {
     /// <summary> Custom VisualAttribute drawer for <see cref="FoldoutGroupAttribute"/> </summary>
     [Artifice_CustomAttributeDrawer(typeof(FoldoutGroupAttribute))]
-    public class Artifice_CustomAttributeDrawer_FoldoutGroupAttribute : Artifice_CustomAttributeDrawer_BoxGroupAttribute
+    public class Artifice_CustomAttributeDrawer_FoldoutGroupAttribute : Artifice_CustomAttributeDrawer_GroupAttribute
     {
         protected override Type VisualElementType { get; } = typeof(Artifice_VisualElement_FoldoutGroup);
     }