+++ Editor/Artifice_CustomAttributeDrawers/CustomAttributeDrawers_Groups/Artifice_CustomAttributeDrawer_FoldoutGroupBeginAttribute.cs
@@ -1,0 +1,10 @@
+using ArtificeToolkit.Attributes;
+
+namespace ArtificeToolkit.Editor.Artifice_CustomAttributeDrawers.CustomAttributeDrawers_Groups
+{
+    [Artifice_CustomAttributeDrawer(typeof(FoldoutGroupBeginAttribute))]
+    public class Artifice_CustomAttributeDrawer_FoldoutGroupBeginAttribute : Artifice_CustomAttributeDrawer_FoldoutGroupAttribute
+    {
+        protected override bool IsOpenGroupDrawer { get; } = true;
+    }
+}