+++ Editor/Artifice_VisualElements/Artifice_VisualElement_SlidingGroup.cs
@@ -1,5 +1,4 @@
 using ArtificeToolkit.Editor.Artifice_CustomAttributeDrawers.CustomAttributeDrawers_Groups;
-using UnityEngine;
 using UnityEngine.UIElements;
 using UnityEngine.UIElements.Experimental;