+++ Runtime/CustomAttributes/BoxGroupBeginAttribute.cs
@@ -1,0 +1,15 @@
+namespace ArtificeToolkit.Attributes
+{
+    public class BoxGroupBeginAttribute : BoxGroupAttribute
+    {
+        public BoxGroupBeginAttribute() : base("Box Group")
+        {
+        }
+        public BoxGroupBeginAttribute(string groupName) : base(groupName)
+        {
+        }
+        public BoxGroupBeginAttribute(string groupName, GroupColor groupColor) : base(groupName, groupColor)
+        {
+        }
+    }
+}