+++ Runtime/CustomAttributes/GroupEndAttribute.cs
@@ -1,0 +1,6 @@
+using ArtificeToolkit.Attributes;
+using UnityEngine;
+
+public class GroupEndAttribute : CustomAttribute
+{
+}