+++ Runtime/CustomAttributes/HorizontalGroupBeginAttribute.cs
@@ -1,0 +1,9 @@
+namespace ArtificeToolkit.Attributes
+{
+    public class HorizontalGroupBeginAttribute : HorizontalGroupAttribute
+    {
+        public HorizontalGroupBeginAttribute(string groupName) : base(groupName)
+        {
+        }
+    }
+}