+++ Runtime/CustomAttributes/VerticalGroupBeginAttribute.cs
@@ -1,0 +1,13 @@
+namespace ArtificeToolkit.Attributes
+{
+    public class VerticalGroupBeginAttribute : VerticalGroupAttribute
+    {
+        public VerticalGroupBeginAttribute() : base("Vertical Group")
+        {
+            
+        }
+        public VerticalGroupBeginAttribute(string groupName) : base(groupName)
+        {
+        }
+    }
+}